Added code to load US census data

+++ src/main/java/censusanalyser/IndiaCensusCSV.java
@@ -8,13 +8,13 @@ public class IndiaCensusCSV {
     public String state;
 
     @CsvBindByName(column = "Population", required = true)
-    public int population;
+    public double population;
 
     @CsvBindByName(column = "AreaInSqKm", required = true)
-    public int areaInSqKm;
+    public double areaInSqKm;
 
     @CsvBindByName(column = "DensityPerSqKm", required = true)
-    public int densityPerSqKm;
+    public double densityPerSqKm;
 
     @Override
     public String toString() {
